ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
opencv-python-headless==4.8.1.78
mediapipe==0.10.8
numpy==1.26.2
python-multipart==0.0.6
Pillow==10.1.0
requests==2.31.0
moviepy==1.0.3
google-generativeai==0.3.1
pydantic==2.5.2
python-dotenv==1.0.0
stripe==7.9.0
pillow-heif==0.13.1
rembg==2.0.52
protobuf==3.20.3
